+++ sync-for-commerce/src/codatsynccommerce/models/shared/halref.py
@@ -12,5 +12,6 @@ from typing import Optional
 @dataclasses.dataclass
 class HalRef:
     href: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('href'), 'exclude': lambda f: f is None }})
+    r"""Uri hypertext reference."""